Implement session authentication, user authorization, and public blog feed

+++ models/blogModel.js
@@ -1,12 +1,12 @@
 import db from "../config/db.js";
 
-export async function getAllBlogs() {
-  const result = await db.query('SELECT * FROM blogs ORDER BY id ASC;');
+export async function getUserBlogs(userId) {
+  const result = await db.query('SELECT * FROM blogs WHERE user_id = $1', [userId]);
   return result.rows;
 }
 
-export async function createBlog(title, body) {
-    await db.query("INSERT INTO blogs (title, body) VALUES ($1, $2)", [title, body]);
+export async function createBlog(title, body, userId) {
+    await db.query("INSERT INTO blogs (title, body, user_id) VALUES ($1, $2, $3)", [title, body,userId ]);
 }
 
 export async function getBlogById(id) {
@@ -14,10 +14,15 @@ export async function getBlogById(id) {
     return result.rows[0];
 }
 
-export async function updateBlog(id, title, body) {
-    await db.query('UPDATE blogs SET title = $1, body = $2 WHERE id = $3;', [title, body, id]);
+export async function updateBlog(id, title, body, userId) {
+    await db.query('UPDATE blogs SET title = $1, body = $2 WHERE id = $3 AND user_id = $4;', [title, body, id, userId]);
 }
 
-export async function deleteBlog(id) {
-    await db.query('DELETE FROM blogs WHERE id = $1', [id]);
+export async function deleteBlog(id, userId) {
+    await db.query('DELETE FROM blogs WHERE id = $1 AND user_id = $2', [id, userId]);
+}
+
+export async function getAllBlogs() {
+    const result = await db.query("SELECT blogs.*, users.email FROM blogs JOIN users ON blogs.user_id = users.id ORDER BY blogs.id DESC")
+    return result.rows;
 }
